+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 - Indices in `pathway.stdlib.indexing.nearest_neighbors` can now work also on numpy arrays. Previously they only accepted `list[float]`. Working with numpy arrays improves memory efficiency.
 - `pw.io.s3.read` has been optimized to minimize new object requests whenever possible.
+- It is now possible to set the size limit of cache in `pw.udfs.DiskCache`.
 
 ## [0.15.1] - 2024-10-04